CASSANDRA-17445-p6 Move all nodetool commands under the picocli
