+++ src/java/org/apache/cassandra/tools/nodetool/JmxConnect.java
@@ -41,7 +41,7 @@ import picocli.CommandLine.RunLast;
 import picocli.CommandLine.Spec;
 
 import static java.lang.Integer.parseInt;
-import static org.apache.cassandra.tools.NodeToolV2.lastExecutableSubcommandWithSameParent;
+import static org.apache.cassandra.tools.NodeTool.lastExecutableSubcommandWithSameParent;
 import static org.apache.commons.lang3.StringUtils.EMPTY;
 import static org.apache.commons.lang3.StringUtils.isEmpty;
 import static org.apache.commons.lang3.StringUtils.isNotEmpty;
@@ -58,19 +58,19 @@ public class JmxConnect extends AbstractCommand implements AutoCloseable
     @Spec
     protected CommandSpec spec; // injected by picocli
 
-    @Option(names = { "-h", "--host" }, description = "Node hostname or ip address")
+    @Option(names = { "-h", "--host" }, description = "Node hostname or ip address", arity = "0..1")
     public String host = "127.0.0.1";
 
-    @Option(names = { "-p", "--port" }, description = "Remote jmx agent port number")
+    @Option(names = { "-p", "--port" }, description = "Remote jmx agent port number", arity = "0..1")
     public String port = "7199";
 
-    @Option(names = { "-u", "--username" }, description = "Remote jmx agent username")
+    @Option(names = { "-u", "--username" }, description = "Remote jmx agent username", arity = "0..1")
     public String username = EMPTY;
 
-    @Option(names = { "-pw", "--password" }, description = "Remote jmx agent password")
+    @Option(names = { "-pw", "--password" }, description = "Remote jmx agent password", arity = "0..1")
     public String password = EMPTY;
 
-    @Option(names = { "-pwf", "--password-file" }, description = "Path to the JMX password file")
+    @Option(names = { "-pwf", "--password-file" }, description = "Path to the JMX password file", arity = "0..1")
     public String passwordFilePath = EMPTY;
 
     @Inject
