+++ src/java/org/apache/cassandra/tools/nodetool/TopLevelCommand.java
@@ -131,6 +131,7 @@ import static org.apache.cassandra.tools.nodetool.Help.printTopCommandUsage;
                          ResetLocalSchema.class,
                          ResumeHandoff.class,
                          Ring.class,
+                         Scrub.class,
                          SetAuthCacheConfig.class,
                          SetBatchlogReplayThrottle.class,
                          SetCacheCapacity.class,
@@ -150,6 +151,7 @@ import static org.apache.cassandra.tools.nodetool.Help.printTopCommandUsage;
                          SetStreamThroughput.class,
                          SetTimeout.class,
                          SetTraceProbability.class,
+                         Sjk.class,
                          Snapshot.class,
                          Status.class,
                          StatusAutoCompaction.class,
@@ -159,7 +161,6 @@ import static org.apache.cassandra.tools.nodetool.Help.printTopCommandUsage;
                          StatusHandoff.class,
                          Stop.class,
                          StopDaemon.class,
-                         Sjk.class,
                          TableHistograms.class,
                          TableStats.class,
                          TopPartitions.class,
